feat：升级loger_viewer，移除无用脚本
