+++ src/chat/chat_loop/heartFC_chat.py
@@ -20,7 +20,7 @@ from src.person_info.person_info import get_person_info_manager
 from src.plugin_system.base.component_types import ActionInfo, ChatMode
 from src.plugin_system.apis import generator_api, send_api, message_api
 from src.chat.willing.willing_manager import get_willing_manager
-from src.chat.mai_thinking.mai_think import mai_thinking_manager
+from src.mais4u.mai_think import mai_thinking_manager
 from maim_message.message_base import GroupInfo
 from src.mais4u.constant_s4u import ENABLE_S4U
