+++ src/mais4u/mai_think.py
